+++ playground/AudioDSP/Boost/Boost.ino
@@ -10,7 +10,6 @@
   Based on OpenMusicLabs previous works.
   pedalshield_uno_booster.ino: pressing the pushbutton_1 or 2 turns the volume up or down.
 
-
  */
 
 #include <AudioDspDriver.h>  // main AudioDSP controller class
@@ -25,8 +24,12 @@ void loop() {
   dsp_driver.process_controls();
 }
 
-// volume transformer
-int transformer(int input, int pb_level) {
+/* volume transformer
+ * the input signal is 16 bit signed (values from -32768 to +32768, INT16_MIN to INT16_MAX)
+ * pb_level is 0-1024 with midpoint 512
+ * the output/transformed value is scaled to the |pb_level| value
+ */
+int16_t transformer(int16_t input, int pb_level) {
   return map(input, 0, 1024, 0, pb_level);
 }
